Refactor `util`'s alias
