--- containers/Home/Achievement/util/index.ts
@@ -1,1 +1,0 @@
-export { useCount } from './countDown'
