+++ global.d.ts
@@ -1,0 +1,5 @@
+declare global {
+    interface Window {
+        gtag: Gtag.Gtag
+    }
+}
