+++ hooks/index.ts
@@ -1,3 +1,4 @@
 export { useIsBrowser } from './useIsBrowser'
 export { useIsPrintMode } from './useIsPrintMode'
 export { useIsMobile } from './useIsMobile'
+export { useCount } from './useCount'
