+++ next.config.js
@@ -9,11 +9,12 @@ module.exports = {
                 ...config.resolve,
                 alias: {
                     ...config.resolve.alias,
+                    '@api': path.resolve((__dirname, 'api/')),
                     '@components': path.resolve((__dirname, 'components/')),
                     '@containers': path.resolve((__dirname, 'containers/')),
-                    '@enum': path.resolve((__dirname, 'enum/')),
                     '@context': path.resolve((__dirname, 'context/')),
-                    '@api': path.resolve((__dirname, 'api/')),
+                    '@enum': path.resolve((__dirname, 'enum/')),
+                    '@hooks': path.resolve((__dirname, 'hooks/')),
                     '@utils': path.resolve((__dirname, 'utils/')),
                 },
             },
