+++ package.json
@@ -10,7 +10,6 @@
         "format": "prettier -w ."
     },
     "dependencies": {
-        "@types/gtag.js": "^0.0.7",
         "axios": "^0.21.4",
         "lodash": "^4.17.21",
         "moment": "^2.29.1",
@@ -22,6 +21,7 @@
         "styled-components": "^5.3.0"
     },
     "devDependencies": {
+        "@types/gtag.js": "^0.0.7",
         "@types/lodash": "^4.14.172",
         "@types/react": "17.0.16",
         "@types/styled-components": "^5.1.12",
