+++ pages/_app.tsx
@@ -1,20 +1,18 @@
 import { ThemeProvider as CustomThemeProvider } from 'context/ThemeContext'
 import { useBrowserEffect } from 'hooks/useBrowserEffect'
 import type { AppProps } from 'next/app'
-import 'style/global.css'
-import { handlePreventCopyCutContent, handlePreventInspectElement } from 'utils/page'
 import { useRouter } from 'next/router'
-import * as ga from 'utils/ga'
 import { useEffect } from 'react'
+import { handlePreventCopyCutContent, handlePreventInspectElement } from '@utils/page'
+import { pageview } from '@utils/google-analytic'
+import 'style/global.css'
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
     useBrowserEffect(() => {
         if (process.env.NODE_ENV === 'production') {
-            ;(function () {
-                const self = document
-                handlePreventCopyCutContent(self)
-                handlePreventInspectElement(self)
-            })()
+            const self = document
+            handlePreventCopyCutContent(self)
+            handlePreventInspectElement(self)
         }
     }, [])
 
@@ -22,7 +20,7 @@ const MyApp = ({ Component, pageProps }: AppProps) => {
 
     useEffect(() => {
         const handleRouteChange = (url: URL) => {
-            ga.pageview(url)
+            pageview(url)
         }
         router.events.on('routeChangeComplete', handleRouteChange)
